docs: delete unused badge (#222)
It seems that there is an unused badge in this repository. This badge
was introduced in
https://github.com/actions/create-github-app-token/pull/70, but after
some trial and error, it was removed from the README
(https://github.com/actions/create-github-app-token/pull/70/commits/f28f8958a77a5e1e1f036ad308ffa00c25948a97
in the PR). However, the badge itself was not deleted.

Therefore, this badge appears to be unnecessary. This patch removes it.