Update `actions/checkout` docs
Updates the `actions/checkout` doc section to demonstrate how to push a commit using the generated token. 

My experience was that I came here because I needed to push with a app token so my CI would run after. 

So an example of how to do that would have been more helpful then a example on how to prevent using this token to push. 

Closing my other PR as I think this better addresses what I came to the docs looking for.